Unpun SingleSelect and Selectable events
Both `SingleSelect` and `Selectable` fire `selected` and `deselected`
events, making event capture on a collection more difficult, as you get
twice the events you want, with different parameters.

Update Selectable's `selected` and `deselected` events to provide the
model as its first parameter (like how the collection `selected` event
behaved previously).

Change `SingleSelect` to fire `select:one` and `deselect:one` events
instead of the `selected` and `deselected` events. This is roughly
similar to the namespacing of the `MultiSelect` and so maintains a
foolish consistency.

Also:

* Update the tests
* Update the Readme documentation