Add support for a model which belongs to more than one collection.
If a selectable model belongs to more than one selectable collection then only the initial collection will have it's select() method called.

Also, if a model is initially added to a non-selectable collection, and then later added to a selectable collection the 2nd collection will not have it's select() method called, and a error will be thrown as the initial collection will not have a select() method.